refactor of the be_well_formed matcher
