+++ spec/quality_spec.rb
@@ -1,50 +1,13 @@
 require "spec_helper"
 
 describe "The application itself" do
-  def check_for_tab_characters(filename)
-    failing_lines = []
-    File.readlines(filename).each_with_index do |line,number|
-      failing_lines << number + 1 if line =~ /\t/
-    end
-
-    unless failing_lines.empty?
-      "#{filename} has tab characters on lines #{failing_lines.join(', ')}"
-    end
-  end
-
-  def check_for_extra_spaces(filename)
-    failing_lines = []
-    File.readlines(filename).each_with_index do |line,number|
-      next if line =~ /^\s+#.*\s+\n$/
-      failing_lines << number + 1 if line =~ /\s+\n$/
-    end
-
-    unless failing_lines.empty?
-      "#{filename} has spaces on the EOL on lines #{failing_lines.join(', ')}"
-    end
-  end
-
-  RSpec::Matchers.define :be_well_formed do
-    failure_message_for_should do |actual|
-      actual.join("\n")
-    end
-
-    match do |actual|
-      actual.empty?
-    end
-  end
-
   it "has no malformed whitespace" do
-    error_messages = []
     Dir.chdir(File.expand_path("../..", __FILE__)) do
-      `git ls-files`.split("\n").each do |filename|
+      files = `git ls-files`.split("\n").find_all do |filename|
         ignore = %w{public/javascripts/rails.js public/javascripts/jquery.js features/support/env.rb features/step_definitions/web_steps.rb lib/tasks/cucumber.rake}
-        next if ignore.include?(filename)
-        next if filename =~ /\.gitmodules|fixtures|\.png$|\.jp(e)g$|\.gif$|/
-        error_messages << check_for_tab_characters(filename)
-        error_messages << check_for_extra_spaces(filename)
+        !ignore.include?(filename) && filename !~ /\.gitmodules|fixtures|\.png$|\.jp(e)g$|\.gif$/
       end
+      files.should be_well_formed
     end
-    error_messages.compact.should be_well_formed
   end
 end
